+++ sdk/core/azure-core-test/src/test_proxy_manager.cpp
@@ -247,6 +247,7 @@ void TestProxyManager::SetProxySanitizer()
   addSanitizer(SanitizerType::Uri, storageSasSignatureRegex, "sassig");
   addSanitizer(SanitizerType::Header, storageSasSignatureRegex, "sassig", "x-ms-copy-source");
   addSanitizer(SanitizerType::Header, storageSasSignatureRegex, "sassig", "x-ms-rename-source");
+  addSanitizer(SanitizerType::Header, "(?<auth>.+)", "auth", "x-ms-copy-source-authorization");
   const std::string storageUserDelegationKeyRegex
       = "\\u003CValue\\u003E(?<userdelegationkey>[a-zA-Z0-9\\/=+]+).*\\u003C\\/"
         "UserDelegationKey\\u003E";