+++ sdk/eventhubs/azure-messaging-eventhubs/samples/basic-operations/create_producer.cpp
@@ -27,6 +27,7 @@ int main()
   if (eventhubName == nullptr)
   {
     std::cerr << "Missing environment variable EVENTHUB_NAME" << std::endl;
+    return 1;
   }
 
   Azure::Messaging::EventHubs::ProducerClient producerClient(