+++ sdk/storage/test-resources-post.ps1
@@ -12,6 +12,8 @@ New-AzStorageEncryptionScope -ResourceGroupName $ResourceGroupName -StorageAccou
 
 Enable-AzStorageBlobDeleteRetentionPolicy -ResourceGroupName $ResourceGroupName -StorageAccountName $DeploymentOutputs['DATALAKE_ACCOUNT_NAME'] -RetentionDays 7
 
+Enable-AzStorageBlobLastAccessTimeTracking -ResourceGroupName $ResourceGroupName -StorageAccountName $DeploymentOutputs['ACCOUNT_NAME'] -PassThru
+
 # This script is used to wait until XCache is refreshed for the service properties (30s), and role assignment takes effect (300s).
 
 Start-Sleep -s 300