+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugs Fixed
 
 - Fixed a bug where `DataLakeDirectoryClient::ListPaths` and `DataLakeFileSystemClient::ListPaths` cannot list the second page and always fail with `std::bad_function_call`.
+- Fixed a bug where `DataLakePathClient::SetAccessControlListRecursive` crashes when operating on the second page.
 
 ### Other Changes