+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Features Added
 
+- TenantId can now be discovered through the service challenge response, when using a TokenCredential for authorization.
+    - A new property is now available on `QueueClientOptions` called `EnableTenantDiscovery`. If set to `true`, the client will attempt an initial unauthorized request to the service to prompt a challenge containing the tenantId hint.
+
 ### Breaking Changes
 
 ### Bugs Fixed